pyproject.toml
[project]
name = "pytorch-project-template"
version = "4.0.0"
description = ""
authors = [{ name = "Junya Morioka", email = "mjun@mjunya.com" }]
dependencies = [
    "torch==2.2.1+cu121",
    "torchvision==0.17.1+cu121",
    "timm>=0.9.16",
    "kornia>=0.7.2",
    "mlflow>=2.15.1",
    "torchmetrics>=1.0.3",
    "natsort>=8.4.0",
    "tabulate>=0.9.0",
    "tqdm>=4.64.1",
    "matplotlib>=3.8.4",
    "omegaconf>=2.3.0",
    "opencv-python>=4.9.0.80",
    "python-dotenv>=1.0.1",
    "torchinfo>=1.8.0",
    "pynvml>=11.5.0",
    "psutil>=5.9.8",
    "torch-cpp @ https://github.com/mjun0812/torch_cpp/releases/download/v1.0.5/torch_cpp-1.0.5+torch221cu122-cp311-cp311-linux_x86_64.whl",
    "einops>=0.8.0",
    "loguru>=0.7.2",
    "pandas>=2.2.2",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["ruff>=0.4.2"]
virtual = true

[[tool.rye.sources]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pyright]
exclude = ["dataset", "result"]

[tool.ruff]
exclude = [".git", ".ruff_cache", ".venv", ".vscode", "dataset", "result"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # flake8
    "I", # isort
]
ignore = [
    "G004",   # `logging-f-string` : loggingでf-stringを使いたいので無視する
    "COM812",
    "COM819",
    "D100",
    "D203",
    "D213",
    "D300",
    "D212",
    "D400",
    "D415",
    "E111",
    "E114",
    "E117",
    "ISC001",
    "ISC002",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "W191",
    "B007",
    "B905",
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pydocstyle]
convention = "google"
